* updated swagger * updated service * handled exception

+++ src/main/java/jp/co/axa/apidemo/service/EmployeeServiceImpl.java
@@ -1,13 +1,16 @@
 package jp.co.axa.apidemo.service;
 
+import jp.co.axa.apidemo.constants.ApiCodes;
+import jp.co.axa.apidemo.constants.Constants;
 import jp.co.axa.apidemo.entity.Employee;
+import jp.co.axa.apidemo.exception.MyCustomException;
 import jp.co.axa.apidemo.model.request.EmployeePostRequest;
 import jp.co.axa.apidemo.model.request.EmployeePutRequest;
 import jp.co.axa.apidemo.repository.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -21,22 +24,20 @@ public class EmployeeServiceImpl implements EmployeeService {
         this.employeeRepository = employeeRepository;
     }
 
-    public List<Employee> getEmployees(Long employeeId) {
-
-        List employeeList = new ArrayList<>();
-        if (employeeId != null && employeeId > 0) {
-            Optional<Employee> employee = employeeRepository.findById(employeeId);
-            employeeList.add(employee);
-        } else {
-            employeeList = employeeRepository.findAll();
+    public Employee getEmployee(Long employeeId) {
+        Optional<Employee> optEmp = employeeRepository.findById(employeeId);
+        if (!optEmp.isPresent()) {
+            throw new MyCustomException(Constants.EMPLOYEE_NOT_FOUND, ApiCodes.NOT_FOUND, HttpStatus.NOT_FOUND);
         }
-        return employeeList;
+        return optEmp.get();
     }
 
+    public List<Employee> getEmployees() {
+        return employeeRepository.findAll();
+    }
 
     public Employee saveEmployee(EmployeePostRequest postEmployee) {
         Employee employee = new Employee();
-        employee.setId(postEmployee.getId());
         employee.setName(postEmployee.getName());
         employee.setDepartment(postEmployee.getDepartment());
         employee.setSalary(postEmployee.getSalary());
@@ -44,18 +45,27 @@ public class EmployeeServiceImpl implements EmployeeService {
     }
 
     public void deleteEmployee(Long employeeId) {
-        employeeRepository.deleteById(employeeId);
+        if (getEmployee(employeeId) != null) {
+            employeeRepository.deleteById(employeeId);
+        }
     }
 
-    public Employee updateEmployee(EmployeePutRequest putEmployee) {
+    public Employee updateEmployee(EmployeePutRequest putEmployee, Long employeeId) {
 
-        Employee employee = new Employee();
-        employee.setId(putEmployee.getId());
-        employee.setName(putEmployee.getName());
-        employee.setDepartment(putEmployee.getDepartment());
-        employee.setSalary(putEmployee.getSalary());
-
-        return employeeRepository.save(employee);
-
+        Employee currentEmployee = getEmployee(employeeId);
+        if (currentEmployee != null) {
+            currentEmployee.setId(employeeId);
+            if (putEmployee.getName() != null) {
+                currentEmployee.setName(putEmployee.getName());
+            }
+            if (putEmployee.getDepartment() != null) {
+                currentEmployee.setDepartment(putEmployee.getDepartment());
+            }
+            if (putEmployee.getSalary() != null) {
+                currentEmployee.setSalary(putEmployee.getSalary());
+            }
+            currentEmployee = employeeRepository.save(currentEmployee);
+        }
+        return currentEmployee;
     }
 }
